Font and icon size adjusted
