+++ pages/styles/globals.scss
@@ -14,6 +14,12 @@
     #77aceb 12.18%,
     #1d428a 106.22%
   );
+
+  font-size: 18px;
+
+  @media only screen and (max-width: 800px) {
+    font-size: 14px;
+  }
 }
 
 html,
