+++ src/pages/Home/About/AboutVideo/AboutVideo.module.scss
@@ -1,0 +1,61 @@
+.About {
+  align-items: center;
+  background-color: var(--conun-bg-blue);
+  padding: 200px 200px;
+
+  & .AboutCardsSection {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: space-between;
+    width: 100%;
+
+    & .AboutCard {
+      background: linear-gradient(
+        180deg,
+        rgba(255, 255, 255, 0) 0%,
+        #fff 100%
+      );
+      border: 3px solid #fff;
+      display: flex;
+      flex-direction: column;
+      height: 600px;
+      justify-content: space-between;
+      width: 32%;
+    }
+  }
+
+  & .VideoContainer {
+    height: 720px;
+    margin-top: 200px;
+    padding: 100px;
+    width: 100%;
+  }
+}
+
+@media only screen and (max-width: 800px) {
+  .About {
+    align-items: center;
+    background-color: var(--conun-bg-blue);
+    padding: 100px 20px;
+    & .AboutCardsSection {
+      display: flex;
+      flex-direction: column;
+      gap: 20px;
+      justify-content: space-between;
+      width: 100%;
+
+      & .AboutCard {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+      }
+    }
+    & .VideoContainer {
+      height: 200px;
+      padding: 0 20px;
+      @media (min-width: 768px) and (max-width: 1024px) {
+        height: 500px;
+      }
+    }
+  }
+}
