+++ src/pages/Home/About/ConunNews/ConunNews.module.scss
@@ -6,6 +6,7 @@
   flex-direction: column;
   font-size: 40px;
   padding: 0 200px;
+  text-align: center;
   width: 100%;
 
   & .Community {
