+++ src/pages/Home/About/RoadMapSection/RoadMapSection.module.scss
@@ -50,8 +50,9 @@
 
     & .Year {
       align-items: center;
-      background-color: white;
+      border: 3px solid #fff;
       border-radius: 50%;
+      color: #77aceb;
       display: flex;
       height: 80px;
       justify-content: center;
@@ -65,6 +66,10 @@
 
       &:hover {
         cursor: pointer;
+        transform: scale(1.3);
+      }
+      &:active {
+        transform: scale(1.3);
       }
     }
     & .ArchLine {
