+++ src/pages/Home/index.tsx
@@ -1,5 +1,6 @@
 import Main from "./Main";
 import AboutIntro from "./About/AboutIntro";
+import AboutVideo from "./About/AboutVideo";
 import UseCaseSection from "./About/UseCaseSection";
 import Product from "./About/Product";
 import RoadMap from "./About/RoadMapSection";
@@ -11,6 +12,7 @@ function Home() {
     <div className={styles.Home}>
       <Main />
       <AboutIntro />
+      <AboutVideo />
       <Product />
       <UseCaseSection />
       <RoadMap />
